third_party/blink/web_tests/external/wpt/css/css-values/minmax-angle-serialize-expected.txt

This is a testharness.js-based test.
FAIL 'rotate(min(90deg))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_equals: 'rotate(min(90deg))' and 'rotate(calc(90deg))' should serialize the same in specified values. expected "rotate(calc(90deg))" but got "rotate(min(90deg))"
PASS 'rotate(min(90deg))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'.
FAIL 'rotate(min(.25turn))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_equals: 'rotate(min(.25turn))' and 'rotate(calc(90deg))' should serialize the same in specified values. expected "rotate(calc(90deg))" but got "rotate(min(0.25turn))"
PASS 'rotate(min(.25turn))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'.
FAIL 'rotate(min(100grad))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_equals: 'rotate(min(100grad))' and 'rotate(calc(90deg))' should serialize the same in specified values. expected "rotate(calc(90deg))" but got "rotate(min(100grad))"
PASS 'rotate(min(100grad))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'.
FAIL 'rotate(max(90deg))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_equals: 'rotate(max(90deg))' and 'rotate(calc(90deg))' should serialize the same in specified values. expected "rotate(calc(90deg))" but got "rotate(max(90deg))"
PASS 'rotate(max(90deg))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'.
FAIL 'rotate(max(.25turn))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_equals: 'rotate(max(.25turn))' and 'rotate(calc(90deg))' should serialize the same in specified values. expected "rotate(calc(90deg))" but got "rotate(max(0.25turn))"
PASS 'rotate(max(.25turn))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'.
FAIL 'rotate(max(100grad))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_equals: 'rotate(max(100grad))' and 'rotate(calc(90deg))' should serialize the same in specified values. expected "rotate(calc(90deg))" but got "rotate(max(100grad))"
PASS 'rotate(max(100grad))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'.
FAIL 'rotate(min(90deg, 92deg, 93deg))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_equals: 'rotate(min(90deg, 92deg, 93deg))' and 'rotate(calc(90deg))' should serialize the same in specified values. expected "rotate(calc(90deg))" but got "rotate(min(90deg, 92deg, 93deg))"
PASS 'rotate(min(90deg, 92deg, 93deg))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'.
FAIL 'rotate(min(93deg, 92deg, 90deg))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_equals: 'rotate(min(93deg, 92deg, 90deg))' and 'rotate(calc(90deg))' should serialize the same in specified values. expected "rotate(calc(90deg))" but got "rotate(min(93deg, 92deg, 90deg))"
PASS 'rotate(min(93deg, 92deg, 90deg))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'.
FAIL 'rotate(min(90deg, 1.58rad, 0.25turn))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_equals: 'rotate(min(90deg, 1.58rad, 0.25turn))' and 'rotate(calc(90deg))' should serialize the same in specified values. expected "rotate(calc(90deg))" but got "rotate(min(90deg, 1.58rad, 0.25turn))"
PASS 'rotate(min(90deg, 1.58rad, 0.25turn))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'.
FAIL 'rotate(min(0.25turn, 1.58rad, 90deg))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_equals: 'rotate(min(0.25turn, 1.58rad, 90deg))' and 'rotate(calc(90deg))' should serialize the same in specified values. expected "rotate(calc(90deg))" but got "rotate(min(0.25turn, 1.58rad, 90deg))"
PASS 'rotate(min(0.25turn, 1.58rad, 90deg))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'.
FAIL 'rotate(max(81deg, 82deg, 90deg))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_equals: 'rotate(max(81deg, 82deg, 90deg))' and 'rotate(calc(90deg))' should serialize the same in specified values. expected "rotate(calc(90deg))" but got "rotate(max(81deg, 82deg, 90deg))"
PASS 'rotate(max(81deg, 82deg, 90deg))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'.
FAIL 'rotate(max(83deg, 82deg, 90deg))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_equals: 'rotate(max(83deg, 82deg, 90deg))' and 'rotate(calc(90deg))' should serialize the same in specified values. expected "rotate(calc(90deg))" but got "rotate(max(83deg, 82deg, 90deg))"
PASS 'rotate(max(83deg, 82deg, 90deg))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'.
FAIL 'rotate(max(90deg, 1.57rad, 0.25turn))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_equals: 'rotate(max(90deg, 1.57rad, 0.25turn))' and 'rotate(calc(90deg))' should serialize the same in specified values. expected "rotate(calc(90deg))" but got "rotate(max(90deg, 1.57rad, 0.25turn))"
PASS 'rotate(max(90deg, 1.57rad, 0.25turn))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'.
FAIL 'rotate(max(0.25turn, 1.57rad, 90deg))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_equals: 'rotate(max(0.25turn, 1.57rad, 90deg))' and 'rotate(calc(90deg))' should serialize the same in specified values. expected "rotate(calc(90deg))" but got "rotate(max(0.25turn, 1.57rad, 90deg))"
PASS 'rotate(max(0.25turn, 1.57rad, 90deg))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'.
FAIL 'rotate(calc(min(30deg) + max(60deg)))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_equals: 'rotate(calc(min(30deg) + max(60deg)))' and 'rotate(calc(90deg))' should serialize the same in specified values. expected "rotate(calc(90deg))" but got "rotate(calc(min(30deg) + max(60deg)))"
PASS 'rotate(calc(min(30deg) + max(60deg)))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'.
FAIL 'rotate(calc(50grad + min(45deg)))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_equals: 'rotate(calc(50grad + min(45deg)))' and 'rotate(calc(90deg))' should serialize the same in specified values. expected "rotate(calc(90deg))" but got "rotate(calc(50grad + min(45deg)))"
PASS 'rotate(calc(50grad + min(45deg)))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'.
FAIL 'rotate(calc(min(45deg) + 50grad))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_equals: 'rotate(calc(min(45deg) + 50grad))' and 'rotate(calc(90deg))' should serialize the same in specified values. expected "rotate(calc(90deg))" but got "rotate(calc(min(45deg) + 50grad))"
PASS 'rotate(calc(min(45deg) + 50grad))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'.
FAIL 'rotate(calc(50grad + max(45deg)))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_equals: 'rotate(calc(50grad + max(45deg)))' and 'rotate(calc(90deg))' should serialize the same in specified values. expected "rotate(calc(90deg))" but got "rotate(calc(50grad + max(45deg)))"
PASS 'rotate(calc(50grad + max(45deg)))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'.
FAIL 'rotate(calc(max(45deg) + 50grad))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_equals: 'rotate(calc(max(45deg) + 50grad))' and 'rotate(calc(90deg))' should serialize the same in specified values. expected "rotate(calc(90deg))" but got "rotate(calc(max(45deg) + 50grad))"
PASS 'rotate(calc(max(45deg) + 50grad))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'.
Harness: the test ran to completion.
